backend/requirement.txt
torch
Pillow
numpy
opencv-python
scikit-learn
numpy
dlib
opencv-python
ultralytics
albumentations
flask
segmentation_models_pytorch
transformers
